add rollup-svg, move Dashboard components out of page
