+++ packages/react-core/package.json
@@ -63,6 +63,7 @@
     "rimraf": "^3.0.2",
     "rollup": "^3.21.5",
     "rollup-plugin-scss": "^4.0.0",
+    "rollup-plugin-svg": "2.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "typescript": "^4.7.4"
   },
