+++ packages/react-core/src/components/DashboardHeader.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
-import { Avatar } from '../Avatar';
-import { Brand } from '../Brand';
-import { Button, ButtonVariant } from '../Button';
-import { Divider } from '../Divider';
-import { Dropdown, DropdownGroup, DropdownItem, DropdownList } from '../Dropdown';
-import { Masthead, MastheadToggle, MastheadMain, MastheadBrand, MastheadContent } from '../Masthead';
-import { MenuToggle } from '../MenuToggle';
-import { Toolbar, ToolbarContent, ToolbarGroup, ToolbarItem } from '../Toolbar';
-import { PageToggleButton } from './PageToggleButton';
+import { Avatar } from './Avatar';
+import { Brand } from './Brand';
+import { Button, ButtonVariant } from './Button';
+import { Divider } from './Divider';
+import { Dropdown, DropdownGroup, DropdownItem, DropdownList } from './Dropdown';
+import { Masthead, MastheadToggle, MastheadMain, MastheadBrand, MastheadContent } from './Masthead';
+import { MenuToggle } from './MenuToggle';
+import { Toolbar, ToolbarContent, ToolbarGroup, ToolbarItem } from './Toolbar';
+import { PageToggleButton } from './Page/PageToggleButton';
 import BarsIcon from '@patternfly/react-icons/dist/esm/icons/bars-icon';
 import CogIcon from '@patternfly/react-icons/dist/esm/icons/cog-icon';
 import HelpIcon from '@patternfly/react-icons/dist/esm/icons/help-icon';
